pyproject.toml
[tool.poetry]
name = "ask-the-sky"
version = "0.1.0"
description = "Bluesky RAG analyzer"
authors = ["Seu Nome <seuemail@exemplo.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pydantic-settings = ">=2.10.1,<3.0.0"
atproto = ">=0.0.62,<0.0.63"
langchain = ">=0.3.27,<0.4.0"
sentence-transformers = ">=5.1.0,<6.0.0"
faiss-cpu = ">=1.12.0,<2.0.0"
langchain-community = ">=0.3.29,<0.4.0"
langchain-google-genai = ">=2.1.10,<3.0.0"
langchain-openai = ">=0.3.33,<0.4.0"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
python-multipart = ">=0.0.20,<0.0.21"
streamlit = ">=1.49.1,<2.0.0"
wordcloud = ">=1.9.4,<2.0.0"
matplotlib = ">=3.10.6,<4.0.0"
nltk = ">=3.9.1,<4.0.0"
langdetect = ">=1.0.9,<2.0.0"
pandas = ">=2.3.2,<3.0.0"
altair = ">=5.5.0,<6.0.0"
python-dotenv = ">=1.0.0,<2.0.0"
pyyaml = ">=6.0.2,<7.0.0"
authlib = ">=1.0.0"
python-jose = {version = ">=3.3.0", extras = ["cryptography"]}
itsdangerous = ">=2.2.0,<3.0.0"
torch = { version = "2.3.1+cpu", source = "pytorch-cpu" }
torchvision = { version = "0.18.1+cpu", source = "pytorch-cpu" }
torchaudio = { version = "2.3.1+cpu", source = "pytorch-cpu" }

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.4.7"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"